add condition to check is command is proper

+++ app/main.py
@@ -2,17 +2,19 @@ import os
 
 
 def move_file(command: str) -> None:
-    *_, in_file, out_file = command.split(" ")
-    current_location = os.path.dirname(in_file)
-    destination = os.path.dirname(out_file)
+    if len(command.split()) == 3 and command.split()[0] == "mv":
+        *_, in_file, out_file = command.split(" ")
 
-    if current_location == destination:
-        os.rename(in_file, out_file)
-    else:
-        if not os.path.exists(destination):
-            os.makedirs(destination)
+        current_location = os.path.dirname(in_file)
+        destination = os.path.dirname(out_file)
 
-        with (open(in_file, "r") as in_file,
-              open(out_file, "w") as out_file):
-            out_file.write(in_file.read())
-        os.remove(in_file.name)
+        if current_location == destination:
+            os.rename(in_file, out_file)
+        else:
+            if not os.path.exists(destination):
+                os.makedirs(destination)
+
+            with (open(in_file, "r") as in_file,
+                  open(out_file, "w") as out_file):
+                out_file.write(in_file.read())
+            os.remove(in_file.name)
